Add a mix task to generate a metaprogramming elixir module from yaml (#1033)
* Chore: Add .spr.yml file

My last auth diff got to be sprawling. I wanted stacked diffs but got a
mess instead. So I'm trying out spr.

https://github.com/ejoffe/spr

I might want to combine that with git stackless or git-stack

commit-id:e9bc2f15

* Add a mix task to generate a metaprogramming elixir module from yaml

This creates a mix task that will take example yaml and from that we can
codegen a more pluggable and uniform resource module. The major idea is
that we take the example yaml parse it into an elixir map, pattern match
that into ast, then optimize the ast even more, finally adding glue
code.

This isn't that yet. However it is getting pretty darn reasonable. This
currently can:

- read yaml
- parse it
- clean it
- Transform that into ast for Map.put
- Clean that into ast that uses the KubeExt.Builder module
- Combine that into a module ast
- Use `Macro.to_string` to generate a new code file

This doesn't yet:

- strip out crds
- strip out yaml configs
- strip out the large resources
- make selectors better
- combine everything into materialize
- name the module correctly
- image config?

commit-id:4bc9527e