Better config system state (#1276)
* Switch to using polymorphic embed for system battery config

This diff changes the config field to be a polymorphic embed with one type of config. The empty config. That empty config is good enough to pass tests right now since the settings modules are just a bunch of defaults right now. There's no UI to change the settings.

* Add a BatteryCoreConfig that will hold core configs.

* Add ControlServer.Batteries.Installer.install_all/1

This method allows installing all of the system batteries provided in a single transaction. Additionally it allows providing the default configs. So seed configs batteries come from the inital snapshot.

* Start adding defaults to the configs

* Rework the layout of state.

 - Regulaize naming for SystemState and StateSummary.
 - Create ControlServer.SystemState.Summarizer that takes the current system summary.
 - Add an API for the current system summary. This allows much easier testing of system battery configs and kube state access.

* Add namespace settings from state summary