Add ControlServer Installations in the HomeBase and HomeBaseWeb apps. (#1319)
Summary:
This is a step towards the UI install of our dreams. This adds a VERY basic show and index page. Additionally it adds a new page.

Test Plan:

There are tests added and they all pass.
However this is just a small step on getting to the install UI and APIs needed.