Throw an error when resource is not there in kubernetes resource info (#1374)
Summary:

We want to have a 404 page. So for that to happen the view needs to error out. This adds a get! method that raises an exception if the kube resource is not there. Then we use that in the resource_info view so that later error views can be shown.

Test Plan:
Tried opening a non-existent reource url and it raises the informative exception.