Create a `common_core` application (#1412)
Summary:
There's a need for an application that contains the ecto schemas but doesn't spin up a sql connection.
We also want to make SystemState.StateSummary properly typed with the db types
We also want to be able to use the correct config types in `kube_resources`

So that requires a central application in the umbrella project.

Test Plan:
Everything compiles..... It's been a journey for that.
Mix test clean
Then it boots.