Add the ability to copy postgres credentials to other namespaces (#1420)
Summary:
A little while ago I split the battery-core namespace to play better with bootstrapping and istio. Doing that moved some databases and services using them into different namespaces. That exposed that it's not awesome to have to put your services in the namespace that the database is running. So users would need this ability as well. This diff adds the feature that users would want.

- Add PGCredentialCopy as a holder for those settings in the database
- Adds a migration to put that field in the table
- Adds code in `kube_resources` to copy the values from one namespace to another safely.
- Adds the UI for editing
- Changes the show UI to list the namespaces

Test Plan:

- Edited the control server database to have new copies
- Verified that they are there
- Mix tests pass