fix: retry on mix setup (and seed) (#1975)
Adding some debugging I was able to figure out that mix setup is sometimes erroring out because the user doesn't have permissions to create a table. There's no real way to know if the postgres operator is going to decide to take a break and go to fucking lunch.

So rather than wait some time and then try. We'll just retry a bunch of times with no wait. Downside is we fail the first run of setup sometimes. The good news is it's faster even with a couple of retries.

Test Plan:

I ran this over and over again until it always come up clean with the expected batteries:

`./dev/nuke-clusters.sh && ./dev/bootstrap.sh`