fix: don't record potentially stale resources in the database (#2005)
We were putting the whole resource in the oban table to check if it's still there. However each of the checks are then in memory. It makes no sense to store the list in between for something that we are going to retry again, and depends on timely last results of sync.