feat: allow for choosing which container to tail for logs (#2296)
Summary:
Pods that have more than one container were an issue. We need to choose
the container name. So move the tail logs into per container and allow
choosing which one.

- Choose a default container for tailing
- Fix where the container name gets passed to kuberentes. It's not
  supposed to be on the path.
- Use a table to display containers for now.

Test Plan:

Opened up the postgres-operator pod tried tailing the log of a specific
container.

Opened up the postgres-operator deployment and tried tailing the pod
from that. It redirected to a working log.

Tried tailing the working postgresql stuff and it still works.