build: trim up release build with npm (#3109)
Summary:
The last iteration of the platform builds for phoenix apps included a
step that built the static assets seperately from the release. The
intent was to cache the build, and to make sure that front end teams had
their own playground.

However that wasn't worth the complexity trade offs of building
differently for nix vs dev. So this uses nodejs during the release build
step and links in node_modules.

Test Plan:

CI should build all the things.