feat: most of control server rest api (#3209)
Summary:

Programmers love to be able to poke at things. So give a rest api to the
structs that represent user requests.

This PR adds in:

- Postgres Cluster crud
- Redis Cluster crud
- Jupyter Notebook crud
- Knative Service crud

In doing that I had to change error views and it showed that we had
three different Gettext modules that were half merged. I fixed that.

Not Done:
Battery Install/Edit. Since we need to make this flow better I left it
out for now. Users can click the button first.

Test Plan:
Added tests for all resources exposed.