fix: don't panic on destroy if outputs aren't present (#3311)
Summary:
When starting a cluster if you have an error not all the outputs will be
full. That would cause an error when trying to build the stacks to tear
them down. This guards all the places output is touched with nil checks

In the end this was enough to get destroycluster to work on half created
stacks.

Test Plan:
createcluster -> ctr-c

clean up the lock files (we'll need to automate that)

Then destroycluster works.