feat: add install_id to battery core config (#3542)
* feat: add install_id to battery core config

Summary:
We want to report on the usage for an install and check on it's payment
status. For that we need to know the id. However we're also going to
want that information in home base.

This is the first step towards the whole plan of that. Next step is
hooking up teams and home base seeding.

Test Plan:
bix bootstrap && bix dev

* fix: fix integration test slug location