feat: import the openapi structs for lago
Summary:
Lago or GetLago is a source available billing offering. It's nice that
we can hit the easy button the billing emails. However if they ever are
off the rails, we can self host.

I like that trade off.... Not sure why. /shrug

This is the PR for pulling in their structs for open api. That needed
some openapi generation changes. That code is old an shitty. I'm
sorryish for it.

This contains:

- Allow fileds that are specified as asOf with a type and a
  description.

- Special case some fields for lago that are either enums or the types
  are typedefs for arrays.

- Some code generated and then massaged.

Test Plan:
CI Passes and all code style checks.