fix: turn off tag resolution for ECR on knative
Summary:
Docker tags should be immutable (except for latest type) however in
practice that's not the case. So Knative will pre-determine the sha hash
of each docker container and use that sha as the revision.

However AWS's registry requires tokens for metadata operations even
reading. So turn that off for aws.

Test Plan:
/test locally.