feat: cache building the bi binary into a shared location
Summary:
bi can be used a lot and building it everytime sucks. We also need a
stable path for kubernetes auth helpers to refer to. So lets use the
tag.

Binaries will be built once per commit you're on.

They are cleaned up in the .envrc so should be relatively capped.

Test Plan:
Used it