chore: relicense all our code to Batteries Included License, Version 1.0
Summary:
All code before this was unlicensed and copyright Batteries Included.
From this commit on it will continue to be copyright Batteries Included
but released under the Batteries Included License, Version 1.0.

- Add LICENSE to the repo so that github will pick this up as our root
  license
- Add APACHE-2.0.txt as a copy of the text of the license that is
  referenced in Grant of Future License
- Start changing the README to represent that others will be here with
  license.

We are not ready to accept outside contributions until we have a CLA set
up. TODO.

This does not apply to any third party code from open source used or
linked in.

Test Plan:
- CI
- netlify preview deploys