feat: use a wireguard gateway for kind clusters (#333)
So that we can access internal service IPs with Docker Desktop for
Mac/Windows.

Has the nice side effect of removing `testcontainers-go`, which while
nice does bring in a bunch of cruft.

I've been able to test it locally, up to the point of waiting on the
access info. But given bootstrapping is playing up atm this step never
completes. Either way everything looks healthy.

I don't have a Mac handy so someone else will have to verify it works on
mac.

Please note the using wireguard config isn't necessary / or valid on
linux, due to the VPC networks being directly routable from the host.