chore: remove oidc dep (#645)
Summary:
Since it looks like we are going to go the way of oauth2 library we have
too many librarys in that area. Lets clear out the under used code. We
never got to use this an a lighter weith solution is all that's needed
since we only use well known url to get the enpoints. That doesn't need
another dependency.

Test Plan:
CI is clean