fix: don't watch for keycloak coming up on snapshot apply
Summary:

Rather than waiting for keycloak to be up, assume that it's up as soon
as the battery is started. We already catch all errors and exceptions
when trying to get the state or apply. So this should just remove one
place that can race.

This fixes: #649

Test:
Started control server
Started keycloak

Snapshot apply has keycloak actions

Eventually Batteries Included realm is created.