feat: automatically add the admin role on master realm (#800)
Summary:

The master realm doesn't have admin client's created that allow access
to the admin side (and the ui). Instead it has a realm role `admin`. So
this PR makes `UserManager` automatically choose the type of admin to
make a user When we make a user an admin, we're giving keycloak the
permission to treat this user as an admin for all things below
this secuirty. So this gives everything on master realm.

This should fix: #674

Test Plan:
- bix bootstrap
- bix dev
- start the keycloak battery
- wait a few mins for tokens to be created
- create a user on the master realm.
- make them admin
- reset their password
- they can log into the master realm console
- they can see/interact with other realms admin consoles