chore: iterate on build and locale
Summary:

Right now if we try and tail the logs of the control server on
kubernetes it errors out asking about encoding. That log message is a
debug log about phoenix request time to open up the live view socket. So
this is an attempt to remedy that.

- Don't include some plugs in non-dev endpoints
- Set LANG ang LC_ALL early in the process. Don't do that over and over
  again. Simplufy and share it everywhere
- localedev and update-local in addition.

Test Plan:
`bix build-image` for control-server and others
Then let CI take it