feat: add read_only_fields to CommonCore.Ecto.Schema (#899)
Summary:
- Add an annotation `@read_only_fields` that will all the field to be
set
  once on insert, but not after that.
- Add that annotation to lots of secret fields in battery configs
- Apply that annotation to name fields of things that create storage. We
  can't rename a database since that will change the mount name where
  data is.
- Change `CommonCore.Ecto.Schema` new, load, and cast behavior to pass
  the action to changeset.

Note there's a big giant hack in this thing. Ecto uses a new changeset
for each relation. They don't pass the action down into the the
dependent changeset. So we rely on the private metadata field to see if
this record was loaded from the database (in which case we continue on)
or built in memory. In the latter case we assue that read only doesn't
apply. Yes I know it's a hack.

Test Plan:
- Many tests changed
- Updated tests where name can't be changed anymore