Fix email formatting issues (#952)
This fixes #912 by cleaning up the assigns in the email templates and
fetching the static site URL from config rather than hard-coding it.
This lets us use assets from the static site in emails.

I also updated the footer street address, and added the ability to
disable it in specific emails by setting `use HomeBaseWeb, {:email,
street_address: nil}` in the template. I wasn't sure which emails should
have an address and which shouldn't, so I left it in all of them for
now. We can remove it from specific templates in the future as needed.