refactor(docker): streamline image builds (#1902)
The goal here is to have a build and deploy base image that change
slowly. And then platform/pastebin images that change only as needed.

This shrinks our deployable images by ~50MB.

Additionally, it should help CI build times as well as alleviate the
cache pressure we're experiencing as the docker cache takes up a good
chunk of our allotment.

Co-authored-by: Jason Tarasovic <2999202+JTarasovic@users.noreply.github.com>