feat: seed static projects into internal_prod home (#2059)
Description:
- Add HomeBase.Release.do_seed_static_projects
- Add HomeBase.Seed.seed_static_projects
- Add HomeBase.BatteriesInstalls
- Add HomeBase.BatteriesInstalls.list_internal_prod_installations
- Add HomeBase.Projects.StaticProjects

This uses IncludeResource to embed the project snapshots that we want to
always be publicly visible. Then during the seed stage (which is called
on init) we check to see if a stored project snapshot with that same id
exists. If it doesn't then we create it. If it does then nothing
happens. Notice that this will mean that there are no project updates
only insert. So versioning will be by adding a new stored project
snapshot for each version. Deleting will have to come from a ui later.

Test Plan:
- bix l bootstrap && bix l dev
- Create a new project importing, choosing the BI.Example that was
visible.
- The import worked