test: wait longer and earlier for keycloak (#2216)
Description:
Keycloak sso test is the last problem child. So this is an attempt to
get the integration tests to pass on CI well. It first waits for
keycloak's postgres cluster to be up. Then it starts waiting for the
statefulset to be good. I also increased the max_wait_time so that
waiting for the table row to appear can wait longer. For me I can't get
this to fail on my machine any more. However CI was always the fun

Test Plan:
- Ran locally
- CI