test: add the ability to use a specified bi in the start and start local scripts (#2648)
Summary:
This should allow home base int-test to work with changes to bootstrap,
like the one we just had. BI_OVERRIDE_LOC was a hard enough to find env
variable that I picked it. Users won't think they need to set it. But it
should allow this to work.

Test Plan:
/ shrug