ci: fix int-test assertion (#2649)
I've seen failures over the last couple of days like this:

`Expected to find at least 6 visible elements that matched the css
'table tbody tr', but only 3 visible elements were found.`

I'm not sure why we're requiring a minimum of 6 rows. I think we're just
trying to assert we're on the right page.

Co-authored-by: Jason Tarasovic <2999202+JTarasovic@users.noreply.github.com>