fix: wait for metallb to be stable on bootstrap (#2650)
Summary:
MetalLB does the load balancing. I have seen the speaker take a while.
Looking at the timing the controller comes up and assigns an ip before
the speaker is 100% started. So there's a window where the control
server could be up and the ip connections are not ready yet. I think
this could be some of the network disconnected/refused errors we have
seen.

Test Plan:
- CI
- seat of my pants