fix: don't hammer the api server on not-found (#2723)
Summary:
I left in a field by mistake. I was trying to use just one field for the
current retry time. However there was still one place referencing the
old field. That meant that we never grew the base retry time for and
watchers. We hammered the api server on starup.

Test:
- Code review
- Starts everything here